Migrated create_message function, changed default source value, documentation updates
- Migrated create_message function from the to_splunk tool to
splunk_api.py.
- Changed the default 'soruce' value from None to 'lqmt'.
- Updated documentation for more clarity. Starting to better detail how
LQMT interfaces with Splunk and how users should setup Splunk to provide
a safe interface for LQMT.
