+++ lqmt/tools/to_splunk/tool.py
@@ -1,11 +1,9 @@
 import logging
 import requests
-import time
 from requests.packages.urllib3.exceptions import InsecureRequestWarning
 from lqmt.lqm.data import AlertAction
 from lqmt.lqm.tool import Tool
-from lqmt.tools.to_splunk.splunk_api import ApiCaller
-
+from lqmt.tools.to_splunk.splunk_api import ApiCaller, create_message
 
 
 class ToSplunk(Tool):
@@ -25,13 +23,13 @@ class ToSplunk(Tool):
             requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
 
         self.apicaller = ApiCaller(
-                self._config.host,
-                self._config.port,
-                self._config.username,
-                self._config.password,
-                cert_check=self._config.cert_check,
-                source=self._config.source,
-                sourcetype=self._config.sourcetype
+            self._config.host,
+            self._config.port,
+            self._config.username,
+            self._config.password,
+            cert_check=self._config.cert_check,
+            source=self._config.source,
+            sourcetype=self._config.sourcetype
         )
 
     def initialize(self):
@@ -41,23 +39,7 @@ class ToSplunk(Tool):
         """
         Process function. Handles the processing of data for the tool. 
         """
-        self.apicaller.send_message(self.create_message(alert))
-
-    @staticmethod
-    def create_message(alert):
-        """
-        Creates a formatted message for Splunk that contains parsed data from FlexT. Currently returns all non-empty
-        parsed data.
-
-        :param alert: Parsed alert data from FlexT
-        :return: Returns formatted string.
-        """
-        data = alert.getAllFields(dictionary=True, parseEmpty=True)
-        message = "{0} LQMT: ".format(time.asctime())
-        for key, value in data.items():
-            message += "{0}={1} ".format(key, value)
-
-        return message
+        self.apicaller.send_message(create_message(alert))
 
     def commit(self):
         pass
